refactor: navbar
